Refine named_barrir and semaphore for gfx13 (#238)
1. Update the implementation of named_barrier and semaphore to match new interface
- semaphore is enabled with new built-in functions.
- named barrier is switched to asm mode, since new interface is incompatible with current FFM implementation
2. Enable all disabled case in convolution tests, include cases in conv_op and conv_device_wavegroup
3. Fix build error related with amd_wave_read_first_lane
 - Add utility in client_example cmake to avoid build error on gfx9

+++ projects/composablekernel/include/ck/tensor_operation/gpu/device/impl/device_grouped_conv_fwd_multiple_d_xdl_large_tensor_cshuffle.hpp
@@ -67,10 +67,10 @@ __global__ void
     const long_index_t e_group_offset = amd_wave_read_first_lane(
         static_cast<int64_t>(compute_ptr_offset_of_groups.GetEPtrOffset(g_idx)));
 
-    const long_index_t a_n_offset =
-        amd_wave_read_first_lane(compute_ptr_offset_of_n.GetAPtrOffset(n_idx));
-    const long_index_t e_n_offset =
-        amd_wave_read_first_lane(compute_ptr_offset_of_n.GetEPtrOffset(n_idx));
+    const long_index_t a_n_offset = amd_wave_read_first_lane(
+        static_cast<int64_t>(compute_ptr_offset_of_n.GetAPtrOffset(n_idx)));
+    const long_index_t e_n_offset = amd_wave_read_first_lane(
+        static_cast<int64_t>(compute_ptr_offset_of_n.GetEPtrOffset(n_idx)));
 
     index_t left     = 0;
     index_t right    = gemms_count;
